scala.js 1.0.0

+++ js/src/main/scala/no/skytteren/scalatime/scalatime.scala
@@ -5,14 +5,14 @@ package object scalatime {
   implicit class DatePimp(date: Date.type) {
     def now(): Date = {
       val jDate = new scalajs.js.Date()
-      Date(Year(jDate.getFullYear()), Month(jDate.getMonth() + 1), DayOfMonth(jDate.getDate()))
+      Date(Year(jDate.getFullYear().toLong), Month(jDate.getMonth().toShort + 1), DayOfMonth(jDate.getDate().toShort))
     }
   }
 
   implicit class TimePimp(time: Time.type) {
     def now(): Time = {
       val jTime = new scalajs.js.Date()
-      new Time(Hour(jTime.getHours), Minute(jTime.getMinutes), Second(jTime.getSeconds), Millisecond(jTime.getMilliseconds()))
+      new Time(Hour(jTime.getHours.toShort), Minute(jTime.getMinutes.toShort), Second(jTime.getSeconds.toShort), Millisecond(jTime.getMilliseconds().toShort))
     }
   }
 
@@ -20,8 +20,8 @@ package object scalatime {
     def now(): DateTime = {
       val jDateTime = new scalajs.js.Date()
       new DateTime(
-        Year(jDateTime.getFullYear()), Month(jDateTime.getMonth() + 1), DayOfMonth(jDateTime.getDate()),
-        Hour(jDateTime.getHours), Minute(jDateTime.getMinutes), Second(jDateTime.getSeconds), Millisecond(jDateTime.getMilliseconds())
+        Year(jDateTime.getFullYear().toLong), Month(jDateTime.getMonth().toShort + 1), DayOfMonth(jDateTime.getDate().toShort),
+        Hour(jDateTime.getHours.toShort), Minute(jDateTime.getMinutes.toShort), Second(jDateTime.getSeconds.toShort), Millisecond(jDateTime.getMilliseconds().toShort)
       )
     }
   }
